+++ src/main/java/com/ty/attendancesystem/config/TwilioConfig.java
@@ -1,13 +1,18 @@
 package com.ty.attendancesystem.config;
 
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
-@ConfigurationProperties(prefix = "twilio")
 public class TwilioConfig {
+  @Value("twilio.account_sid")
   private String accountSid;
+
+  @Value("twilio.auth_token")
   private String authToken;
+
+  @Value("twilio.trial_number")
   private String trialNumber;
 
   public TwilioConfig(){
