--- src/main/java/com/ty/attendancesystem/model/Admin.java
@@ -1,58 +1,0 @@
-package com.ty.attendancesystem.model;
-
-
-import javax.persistence.*;
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
-import java.util.Date;
-
-@Entity
-@Table(name = "admin")
-public class Admin {
-  @Id
-  @GeneratedValue(strategy = GenerationType.IDENTITY)
-  @Column(name = "id")
-  private Long id;
-
-  @Column(name = "full_name")
-  private String fullName;
-
-  @Column(name = "birthday")
-  private Date birthday;
-
-  @OneToOne(cascade = CascadeType.ALL)
-  @JoinColumn(name = "users_id")
-  private User user;
-
-  public Long getId() {
-    return id;
-  }
-
-  public void setId(Long id) {
-    this.id = id;
-  }
-
-  public String getFullName() {
-    return fullName;
-  }
-
-  public void setFullName(String fullName) {
-    this.fullName = fullName;
-  }
-
-  public Date getBirthday() {
-    return birthday;
-  }
-
-  public void setBirthday(Date birthday) {
-    this.birthday = birthday;
-  }
-
-  public User getUser() {
-    return user;
-  }
-
-  public void setUser(User user) {
-    this.user = user;
-  }
-}
