--- src/main/java/com/ty/attendancesystem/model/Teacher.java
@@ -1,43 +1,0 @@
-package com.ty.attendancesystem.model;
-
-import javax.persistence.*;
-import java.util.Date;
-
-@Entity
-@Table(name = "teacher")
-public class Teacher{
-  @Id
-  @GeneratedValue(strategy = GenerationType.IDENTITY)
-  @Column(name = "id")
-  private Long id;
-
-  @Column(name = "full_name")
-  private String fullName;
-
-  @Column(name = "birthday")
-  private Date birthday;
-
-  public Long getId() {
-    return id;
-  }
-
-  public void setId(Long id) {
-    this.id = id;
-  }
-
-  public String getFullName() {
-    return fullName;
-  }
-
-  public void setFullName(String fullName) {
-    this.fullName = fullName;
-  }
-
-  public Date getBirthday() {
-    return birthday;
-  }
-
-  public void setBirthday(Date birthday) {
-    this.birthday = birthday;
-  }
-}
