--- src/main/java/com/ty/attendancesystem/repository/TeacherRepository.java
@@ -1,8 +1,0 @@
-package com.ty.attendancesystem.repository;
-
-import com.ty.attendancesystem.model.Admin;
-import com.ty.attendancesystem.model.Teacher;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface TeacherRepository extends JpaRepository<Teacher, Long> {
-}
