+++ src/main/resources/application.properties
@@ -19,7 +19,7 @@ spring.messages.encoding=UTF-8
 
 spring.datasource.url=jdbc:postgresql://localhost:5432/attendance_system_db
 spring.datasource.username=postgres
-spring.datasource.password=admin123
+spring.datasource.password=123456
 spring.jpa.database-platform=org.hibernate.dialect.PostgreSQLDialect
 #
 ## Logging
@@ -44,3 +44,9 @@ grokonez.app.jwtExpiration=600000
 twilio.account_sid=AC9ffb1850bb693ff2d4c5cb1487a0ddd8
 twilio.auth_token=c4018c7349a0bcb96b7ea334e882b122
 twilio.trial_number=+16162131749
+#
+## Liquibase
+#
+#spring.liquibase.enabled=true
+#spring.liquibase.change-log=classpath:db/changelog/changelog-master.xml
+#spring.liquibase.drop-first=false
