+++ src/main/resources/db/changelog/changes/changelog-1.1.xml
@@ -1,0 +1,37 @@
+<?xml version="1.1" encoding="UTF-8" standalone="no"?>
+<databaseChangeLog xmlns="http://www.liquibase.org/xml/ns/dbchangelog" xmlns:ext="http://www.liquibase.org/xml/ns/dbchangelog-ext" xmlns:pro="http://www.liquibase.org/xml/ns/pro" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://www.liquibase.org/xml/ns/dbchangelog-ext http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-ext.xsd http://www.liquibase.org/xml/ns/pro http://www.liquibase.org/xml/ns/pro/liquibase-pro-4.1.xsd http://www.liquibase.org/xml/ns/dbchangelog http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-4.1.xsd">
+    <changeSet id="1" author="Tuan">
+        <insert tableName="role">
+            <column name="id" value="1"/>
+            <column name="name" value="ROLE_ADMIN"/>
+        </insert>
+        <insert tableName="role">
+            <column name="id" value="2"/>
+            <column name="name" value="ROLE_TEACHER"/>
+        </insert>
+        <insert tableName="role">
+            <column name="id" value="3"/>
+            <column name="name" value="ROLE_PARENT"/>
+        </insert>
+        <insert tableName="role">
+            <column name="id" value="4"/>
+            <column name="name" value="ROLE_STUDENT"/>
+        </insert>
+    </changeSet>
+    <changeSet id="2" author="Tuan">
+        <insert tableName="users">
+            <column name="username" value="admin" />
+            <column name="password" value="$2y$12$.8t48excE02z8eUnD6YpUeR/fPz0/v3d5yqbTJMBCBlnvaSzDX6aG"/>
+            <column name="email" value="tuantrananhit@gmail.com"/>
+            <column name="phone" value="0377158365"/>
+            <column name="full_name" value="Tran Anh Tuan"/>
+            <column name="birthday" value="1999/12/12"/>
+        </insert>
+    </changeSet>
+    <changeSet id="3" author="Tuan">
+        <insert tableName="users_role">
+            <column name="users_id" value="1"/>
+            <column name="role_id" value="1"/>
+        </insert>
+    </changeSet>
+</databaseChangeLog>
