+++ src/main/resources/db/liquibase.properties
@@ -1,0 +1,6 @@
+driver=org.postgresql.Driver
+url=jdbc:postgresql://localhost:5432/attendance_system_db
+username=postgres
+password=123456
+changeLogFile=src/main/resources/db/changelog/changelog-master.xml
+#outputFile=src/main/resources/db/output.xml
